SRT読込時の再生パフォーマンスを大幅改善
previewPlaybackTick による Home 全体の毎フレーム再レンダーを PreviewSeekBar へ隔離し、字幕キューは ref + 件数 state のみに変更。getEnabled で表示 OFF 時の rAF 字幕処理も完全スキップする。
